Textual and UI changes in material KG generation #1292
blue text should not appear once expand and publish button is clicked. it should just appear only when the KG is in editable mode.
Once the KG is published, the UI moves a lot upwards, removing the tittle, as well as leaving a lot of empty space at the bottom.
the button to close the published material KG is also not functioning